requirements.txt
wsgiref==0.1.2
django-mediagenerator==1.11
django-registration==0.7
Django==1.3
psycopg2==2.4.2
mimeparse>=0.1.3
python-dateutil==1
pystache==0.3.1
django-piston==0.2.2.1
